Adds a few tests
