+++ kifi-backend/modules/common/test/com/keepit/model/NormalizationTest.scala
@@ -1,0 +1,12 @@
+package com.keepit.model
+
+import org.specs2.mutable.Specification
+
+class NormalizationTest extends Specification {
+
+  "Normalization companion object" should {
+    "initialize correctly" in {
+      Normalization.priority.contains(null) === false
+    }
+  }
+}
